requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
debugpy==1.6.2
decorator==5.1.1
entrypoints==0.4
executing==0.9.1
filelock==3.7.1
greenlet==1.1.2
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
newsdataapi==0.1.5
nltk==3.7
numpy==1.23.1
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pika==1.2.1
prompt-toolkit==3.0.30
protobuf==4.21.4
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
regex==2022.7.25
requests==2.28.1
sacremoses==0.0.53
scikit-learn==1.1.2
scipy==1.9.0
sentencepiece==0.1.91
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.39
stack-data==0.3.0
textblob==0.17.1
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing-extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
